requirements.txt
# 文档处理
tika==2.6.0
python-docx==0.8.11
PyPDF2==3.0.1

# 向量化和存储
sentence-transformers==2.2.2
faiss-cpu==1.7.4
SQLite-utils==3.35

# UI
PyQt6==6.5.2

# 工具
watchdog==3.0.0
numpy==1.24.3
pandas==2.0.3